# AIM-OpenAPI.yaml
openapi: 3.0.3
...
info:
  title: Armory Inventory Management APIs
  description: APIs for managing AIM (Armory Inventory Management)
  version: 1.0.0
...
paths:
...
  /inventory:
    get:
      summary: Get all inventories
      operationId: getAllInventories
      tags:
        - Inventory
      parameters:
        - $ref: '#/components/parameters/X-Tenant-ID'  # Repeated for consistency
      responses:
        '200':
          description: A list of Inventories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryDto'
        '500':
          description: Internal server error
#          content:
#            application/json:
#              schema:
#                type: object
#                properties:
#                  error:
#                    type: string
#                    example: "Internal server error"

    # Create an inventory
    post:
      summary: Create a new inventory
      operationId: createInventory
      tags:
        - Inventory
      parameters:
        - $ref: '#/components/parameters/X-Tenant-ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryDto_Id'
      responses:
        '201':
          description: Inventory created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryDto'
        '400':
          description: Bad request - Missing or invalid data
        '409':
          description: Inventory with same serial no is already present.
        '500':
          description: Internal server error
#          content:
#            application/json:
#              schema:
#                type: object
#                properties:
#                  error:
#                    type: string
#                    example: "Internal server error"

  # Get inventory by ID
  /inventory/{inventoryId}:
    get:
      summary: Get inventory by inventoryId
      operationId: getInventoryById
      tags:
        - Inventory
      parameters:
        - name: inventoryId
          in: path
          required: true
          description: ID of the inventory to fetch
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/X-Tenant-ID'
      responses:
        '200':
          description: Inventory details by inventoryId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryDto'
        '404':
          description: Inventory not found
#          content:
#            application/json:
#              schema:
#                type: object
#                properties:
#                  error:
#                    type: string
#                    example: "Inventory not found"
        '500':
          description: Internal server error



    # Update Inventory by Id
    put:
      summary: Update inventory by inventoryId
      operationId: updateInventory
      tags:
        - Inventory
      parameters:
        - name: inventoryId
          in: path
          required: true
          description: ID of the inventory to update
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/X-Tenant-ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryDto_Id'
      responses:
        '200':
          description: Inventory updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryDto'
        '404':
          description: Inventory data not found
#          content:
#            application/json:
#              schema:
#                type: object
#                properties:
#                  error:
#                    type: string
#                    example: "Inventory not found"
        '400':
          description: Bad request - Missing or invalid data

        '500':
          description: Internal server error
#          content:
#            application/json:
#              schema:
#                type: object
#                properties:
#                  error:
#                    type: string
#                    example: "Internal server error"

    delete:
      summary: Delete inventory by inventoryId
      operationId: deleteInventory
      tags:
        - Inventory
      parameters:
        - name: inventoryId
          in: path
          required: true
          description: ID of the inventory to delete
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/X-Tenant-ID'

      responses:
        '200':
          description: Inventory deleted successfully
        '404':
          description: Inventory not found
#          content:
#            application/json:
#              schema:
#                type: object
#                properties:
#                  error:
#                    type: string
#                    example: "Inventory not found"
        '500':
          description: Internal server error


  # Get inventory by serial number
  /inventory/serial/{serialNo}:
    get:
      summary: Get inventory by serial number
      operationId: getInventoryBySerialNo
      tags:
        - Inventory
      parameters:
        - name: serialNo
          in: path
          required: true
          description: Serial number of the inventory to fetch
          schema:
            type: string
        - $ref: '#/components/parameters/X-Tenant-ID'

      responses:
        '200':
          description: Inventory details by serial number
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryDto'
        '404':
          description: Inventory not found
#          content:
#            application/json:
#              schema:
#                type: object
#                properties:
#                  error:
#                    type: string
#                    example: "Inventory not found"
        '500':
          description: Internal server error


  # Get inventories by category ID
  /inventory/category/{categoryId}:
    get:
      summary: Get inventories by categoryId
      operationId: getInventoriesByCategoryId
      tags:
        - Inventory
      parameters:
        - name: categoryId
          in: path
          required: true
          description: Category ID to fetch associated inventories
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/X-Tenant-ID'

      responses:
        '200':
          description: List of inventories for the specified category ID
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryDto'
        '404':
          description: No inventory found for the category
#          content:
#            application/json:
#              schema:
#                type: object
#                properties:
#                  error:
#                    type: string
#                    example: "No inventory found for the category"
        '500':
          description: Internal server error

#  # Get inventory by asset item number
#  /inventory/assetItem/{assetItemNumber}:
#    get:
#      summary: Get inventory by asset item number
#      operationId: getInventoryByAssetItemNumber
#      tags:
#        - Inventory
#      parameters:
#        - name: assetItemNumber
#          in: path
#          required: true
#          description: Asset item number of the inventory to fetch
#          schema:
#            type: string
#        - $ref: '#/components/parameters/X-Tenant-ID'
#
#      responses:
#        '200':
#          description: Inventory details by asset item number
#          content:
#            application/json:
#              schema:
#                $ref: '#/components/schemas/InventoryDto'
#        '404':
#          description: Inventory not found
#          content:
#            application/json:
#              schema:
#                type: object
#                properties:
#                  error:
#                    type: string
#                    example: "Inventory not found"
#=============================================================================================================




#=====================================================COMPONENTS===========================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    X-Tenant-ID:
      name: X-Tenant-ID
      in: header
      description: Tenant identifier for multi-tenancy.
      required: true
      schema:
        type: string
#    Authorization:
#      name: Authorization
#      in: header
#      description: "Bearer token for authentication. Format: `Bearer <token>`"
#      required: true
#      schema:
#        type: string
  schemas:
...
    ManufacturerDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the manufacturer
        name:
          type: string
          description: Name of the manufacturer
        status:
          type: string
          description: Status of the manufacturer (e.g., active, inactive)
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the manufacturer was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the manufacturer was last updated
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryDto'
          description: List of categories associated with the manufacturer.
      required:
        - name

    #Category Schema
    CategoryDto:
      type: object
      properties:
        categoryId:
          type: integer
          format: int64
          description: The unique identifier for the category.
        manufacturers: # Many-to-many relationship with ManufacturerDto
          type: array
          items:
            $ref: '#/components/schemas/ManufacturerDto'
          description: List of manufacturers associated with the category.
        categoryName:
          type: string
          description: The name of the category.
        categoryDetails:
          type: object
          additionalProperties: true
          description: A JSON object containing various details about the category (e.g., model, type, etc.).
          example:
            model: "AR-15"
            type: "Rifle"
            required_training: "Basic"
            required_qualification: "Marksman"
            caliber: "5.56mm"
            length: "36 inches"
            inspection_months: "12"
            maintenance_months: "6"
            description: "A standard assault rifle."

        inventories: # One-to-many association with InventoryDto
          type: array
          items:
            $ref: '#/components/schemas/InventoryDto'
          description: List of inventories associated with the category.
        createdAt:
          type: string
          format: date-time
          description: The timestamp when the category was created.
        updatedAt:
          type: string
          format: date-time
          description: The timestamp when the category was last updated.
      required:
        - categoryName
        - categoryDetails
...
    InventoryDto:
      type: object
      properties:
        inventoryId:
          type: integer
          format: int64
          description: The unique identifier for the inventory item.
#        categoryId:
#          type: integer
#          format: int64
#          description: The unique identifier of the associated category.
        category: # Many-to-one association with CategoryDto
          $ref: '#/components/schemas/CategoryDto'
          description: The associated category for the inventory item.
#        assetItemNumber:
#          type: string
#          description: The unique asset item number for the inventory.
        serialNo:
          type: string
          description: The serial number of the inventory item.
        inventoryDetails:
          type: object
          additionalProperties: true
          description: A JSON object containing varisous details about the inventory item (e.g., lot number, rebuild date, etc.).
          example:
            model: "AQW"
            assetItemNumber: "AS123"
            site: "site1"
            nfa_category: "Rifle"
            description: "Brand new rifle, fully assembled and tested."
        createdAt:
          type: string
          format: date-time
          description: The timestamp when the inventory item was created.
        updatedAt:
          type: string
          format: date-time
          description: The timestamp when the inventory item was last updated.
      required:
        - serialNo
        - categoryId


    #Inventory Dto
    InventoryDto_Id:
      type: object
      properties:
        inventoryId:
          type: integer
          format: int64
          description: The unique identifier for the inventory item.
        categoryId:
          type: integer
          format: int64
          description: The unique identifier of the associated category.
#        category: # Many-to-one association with CategoryDto
#          $ref: '#/components/schemas/CategoryDto'
#          description: The associated category for the inventory item.
        #        assetItemNumber:
        #          type: string
        #          description: The unique asset item number for the inventory.
        serialNo:
          type: string
          description: The serial number of the inventory item.
        inventoryDetails:
          type: object
          additionalProperties: true
          description: A JSON object containing varisous details about the inventory item (e.g., lot number, rebuild date, etc.).
          example:
            model: "AQW"
            assetItemNumber: "AS123"
            site: "site1"
            nfa_category: "Rifle"
            description: "Brand new rifle, fully assembled and tested."
        createdAt:
          type: string
          format: date-time
          description: The timestamp when the inventory item was created.
        updatedAt:
          type: string
          format: date-time
          description: The timestamp when the inventory item was last updated.
      required:
        - serialNo
        - categoryId